+++ src/main/kotlin/io/jawg/geojson/Feature.kt
@@ -6,7 +6,7 @@ data class Feature(
     @JsonInclude(JsonInclude.Include.ALWAYS)
     val geometry: Geometry<*>?,
     @JsonInclude(JsonInclude.Include.ALWAYS)
-    val properties: Map<String, Any>? = null,
+    val properties: Map<String, Any?>? = null,
     var id: String? = null,
     val bbox: BBox? = null
 ) : GeoJsonObject("Feature")