+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.1 2020-01-21
+* `Feature.properties` can have null values as expected by the [RFC](https://tools.ietf.org/html/rfc7946#section-3.2)
+* 3 dimensions `coordinates` don't prevent other properties to be parsed correctly
+
 # 1.2.0 2019-12-04
 * Made `GeoJsonObject` and sub-classes data class
 * Upgrade kotlin to 1.3.61